MODIFY/ 세션 재 설정 및 경로 수정
